feat: allow posting posts and commenting them

+++ domain/interfaces/database.go
@@ -7,7 +7,7 @@ import (
 type IDatabase interface {
 	SavePost(post *models.Post) (models.Post, error)
 	ListPosts() ([]models.Post, error)
-	GetPost(id int) (models.Post, error)
+	GetPost(id int64) (models.Post, error)
 	SaveComment(comment *models.Comment) (models.Comment, error)
-	ListComments(postID int) ([]models.Comment, error)
+	ListComments(postID int64) ([]models.Comment, error)
 }
